// src/features/movies/types/movie.types.ts
export interface Season {
  id: number
  air_date: string | null
  episode_count: number
  name: string
  overview: string
  poster_path: string | null
  season_number: number
  vote_average: number
}

export interface SeasonDetails extends Season {
  episodes: Episode[]
}

export interface Episode {
  id: number
  air_date: string | null
  episode_number: number
  name: string
  overview: string
  runtime: number | null
  season_number: number
  still_path: string | null
  vote_average: number
  vote_count: number
  production_code: string
  show_id: number
  crew: Crew[]
  guest_stars: Cast[]
}
// ...
export interface Cast {
  id: number
  name: string
  character: string
  profile_path: string | null
  order: number
  cast_id: number
  credit_id: string
  adult: boolean
  gender: number
  known_for_department: string
  original_name: string
  popularity: number
}

export interface Crew {
  id: number
  name: string
  job: string
  department: string
  profile_path: string | null
  credit_id: string
  adult: boolean
  gender: number
  known_for_department: string
  original_name: string
  popularity: number
}